Rearranged map import sequence; added dynamic title change
